% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmbrellaManagementAPI.R
\name{addIdentityToPolicy}
\alias{addIdentityToPolicy}
\title{Add an Identity to a directly applied DNS or web policy}
\usage{
addIdentityToPolicy(
  originId,
  policyId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{originId}{originID of the identity to remove from the DNS or web policy}

\item{policyId}{ID of the DNS or web policy to act upon}

\item{organizationId}{Id of the Organization owning the resources}
}
\description{
Policy changes can take up to 20 minutes to take affect globally. Additionally, for DNS policies, TTLs, caching, and session re-use may cause some devices and domains to appear to take longer to update.
}
\concept{UmbrellaManagementAPI}
