% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{basecos}
\alias{basecos}
\title{Cos component of fourier series}
\usage{
basecos(level, t = 1:365/365)
}
\arguments{
\item{level}{level: 1, 2, or 3}

\item{t}{time}
}
\description{
Cos component of fourier series
}
