% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{createApikey}
\alias{createApikey}
\title{Create a management API Key for the customer}
\usage{
createApikey(serviceProviderId, customerId)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerId}{The customerID related to this Customer}
}
\description{
Create a management API Key for the customer
}
\concept{ManagementAPI}
