% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{createCustomer}
\alias{createCustomer}
\title{Create a new Customer}
\usage{
createCustomer(
  serviceProviderId,
  customerName,
  customerType = NULL,
  isTrial = NULL,
  seats,
  streetAddress,
  streetAddress2 = NULL,
  city,
  state,
  countryCode,
  zipCode = NULL,
  packageId,
  dealId = NULL,
  adminEmails,
  ccwDealOwnerEmails = NULL
)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerName}{customer organization name}

\item{customerType}{customer organization type (applicable only for MSSP):   \code{term} = term customer   \code{msla} = msla customer}

\item{isTrial}{applicable only for MSSP with SPLA (MSLA) licensing}

\item{seats}{number of users}

\item{streetAddress}{street address}

\item{streetAddress2}{street address2}

\item{city}{city}

\item{state}{state}

\item{countryCode}{country code}

\item{zipCode}{zip code}

\item{packageId}{package id}

\item{dealId}{deal id}

\item{adminEmails}{array of admin emails}

\item{ccwDealOwnerEmails}{array of CCW deal owner emails}
}
\description{
Create a new Customer
}
\concept{ManagementAPI}
