% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{createCustomerApikeys}
\alias{createCustomerApikeys}
\title{Create a management API Key for the MspCustomer}
\usage{
createCustomerApikeys(mspId, customerId)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerId}{The customerID related to this MspCustomer}
}
\description{
Create a management API Key for the MspCustomer
}
\concept{ManagementAPI}
