% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{createCustomerCustomers}
\alias{createCustomerCustomers}
\title{Create a new MspCustomer}
\usage{
createCustomerCustomers(mspId, customerName, seats)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerName}{customer organization name}

\item{seats}{number of users}
}
\description{
Create a new MspCustomer
}
\concept{ManagementAPI}
