% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DestinationAPIGateway.R
\name{createDestination}
\alias{createDestination}
\title{Add list of destinations to destination list}
\usage{
createDestination(
  destinationListId,
  body = DestinationsList(),
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{destinationListId}{Destination List ID}

\item{organizationId}{Organization ID}

\item{DestinationsList}{List of destinations}
}
\description{
Add list of destinations to destination list
}
\concept{DestinationAPIGateway}
