% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DestinationAPIGateway.R
\name{createDestinationList}
\alias{createDestinationList}
\title{Create a destination list}
\usage{
createDestinationList(
  name,
  access,
  destinations,
  comments = rep("", length(destinations)),
  types = rep("DOMAIN", length(destinations)),
  isGlobal = F,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{name}{New Destination List name}

\item{access}{Access can be allow or block. It defines destinationlist type.}

\item{destinations}{Destination URLs. google.com or google.com/news or 10.10.10.10}

\item{comments}{Comments for each destination.}

\item{types}{Type for each destination. Type can be DOMAIN, URL, IPV4}

\item{isGlobal}{Can be true or false. There will be only one default destination list of type allow or block for an organization.}

\item{organizationId}{Organization ID}
}
\description{
Create a destination list
}
\concept{DestinationAPIGateway}
