% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{createNetwork}
\alias{createNetwork}
\title{Create a new Network. You must first contact support to get your IP range verified before using the POST action to create a network. Reference this API when contacting support.}
\usage{
createNetwork(
  name,
  ipAddress = NULL,
  prefixLength,
  isDynamic,
  status,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{name}{network name}

\item{ipAddress}{network ip address}

\item{prefixLength}{prefix length, must be greater than 28}

\item{isDynamic}{is dynamic ip}

\item{status}{status}

\item{organizationId}{The organizationID related to this Network}
}
\description{
Create a new Network. You must first contact support to get your IP range verified before using the POST action to create a network. Reference this API when contacting support.
}
\concept{ManagementAPI}
