% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmbrellaManagementAPI.R
\name{createNetworkDevice}
\alias{createNetworkDevice}
\title{Create a network device}
\usage{
createNetworkDevice(
  model,
  macAddress,
  name,
  serialNumber,
  tag = NULL,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{model}{The model name of the device. Must be unique to your organization.}

\item{macAddress}{The MAC address of the device (formatted as 12 characters, no hyphens or colons). Must be unique.}

\item{name}{}

\item{serialNumber}{The serial number of the device}

\item{tag}{A text tag that describes the device or this particular origin assigned to the device. Must be unique to your organization.}

\item{organizationId}{Id of the Organization owning the resources}
}
\description{
Create a network device
}
\concept{UmbrellaManagementAPI}
