% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{createSite}
\alias{createSite}
\title{Create a new Site}
\usage{
createSite(name, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{name}{site name}

\item{organizationId}{The organizationID related to this Site}
}
\description{
Create a new Site
}
\concept{ManagementAPI}
