% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteApikey}
\alias{deleteApikey}
\title{Delete a management API key of the customer}
\usage{
deleteApikey(serviceProviderId, customerId, keyId)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerId}{The customerID related to this Customer}

\item{keyId}{The API key id}
}
\description{
Delete a management API key of the customer
}
\concept{ManagementAPI}
