% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteCustomer}
\alias{deleteCustomer}
\title{Delete a Customer}
\usage{
deleteCustomer(serviceProviderId, customerId)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerId}{The customerID related to this Customer}
}
\description{
Delete a Customer
}
\concept{ManagementAPI}
