% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteCustomerApikeys}
\alias{deleteCustomerApikeys}
\title{Delete a management API key of the MspCustomer}
\usage{
deleteCustomerApikeys(mspId, customerId, keyId)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerId}{The customerID related to this MspCustomer}

\item{keyId}{The API key id}
}
\description{
Delete a management API key of the MspCustomer
}
\concept{ManagementAPI}
