% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DestinationAPIGateway.R
\name{deleteDestination}
\alias{deleteDestination}
\title{Delete list of destinations from destination list}
\usage{
deleteDestination(
  destinationListId,
  body = DestinationIdsList(),
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{destinationListId}{Destination List ID}

\item{organizationId}{Organization ID}

\item{DestinationIdsList}{List of destination Id's}
}
\description{
Delete list of destinations from destination list
}
\concept{DestinationAPIGateway}
