% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DestinationAPIGateway.R
\name{deleteDestinationList}
\alias{deleteDestinationList}
\title{Delete destination list}
\usage{
deleteDestinationList(
  destinationListId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{destinationListId}{Destination list ID}

\item{organizationId}{Organization ID}
}
\description{
Delete destination list
}
\concept{DestinationAPIGateway}
