% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteInternaldomain}
\alias{deleteInternaldomain}
\title{Delete an internalDomain}
\usage{
deleteInternaldomain(
  internalDomainId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{internalDomainId}{The internal domain id related to this InternalDomain}

\item{organizationId}{The organization id related to this InternalDomain}
}
\description{
Delete an internalDomain
}
\concept{ManagementAPI}
