% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteInternalnetwork}
\alias{deleteInternalnetwork}
\title{Delete an internalNetwork}
\usage{
deleteInternalnetwork(
  internalNetworkId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{internalNetworkId}{The internal network id related to this InternalNetwork}

\item{organizationId}{The organizationID related to this InternalNetwork}
}
\description{
Delete an internalNetwork
}
\concept{ManagementAPI}
