% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteMspCustomers}
\alias{deleteMspCustomers}
\title{Delete a MspCustomer}
\usage{
deleteMspCustomers(mspId, customerId)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerId}{The customerID related to this MspCustomer}
}
\description{
Delete a MspCustomer
}
\concept{ManagementAPI}
