% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteNetwork}
\alias{deleteNetwork}
\title{Delete a Network}
\usage{
deleteNetwork(networkId, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{networkId}{The network id related to this Network}

\item{organizationId}{The organizationID related to this Network}
}
\description{
Delete a Network
}
\concept{ManagementAPI}
