% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteSite}
\alias{deleteSite}
\title{Delete a Site}
\usage{
deleteSite(siteId, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{siteId}{The siteID related to this Site}

\item{organizationId}{The organizationID related to this Site}
}
\description{
Delete a Site
}
\concept{ManagementAPI}
