% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{deleteUser}
\alias{deleteUser}
\title{Delete a User}
\usage{
deleteUser(userId, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{userId}{The userID related to this User}

\item{organizationId}{The organizationID related to this User}
}
\description{
Delete a User
}
\concept{ManagementAPI}
