% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo.msk}
\alias{demo.msk}
\title{demo.msk
An example availability mask (i.e., the polygon within which simulated reserves can be placed).}
\format{Formal class SpatialPolygons}
\usage{
data(demo.msk)
}
\description{
demo.msk
An example availability mask (i.e., the polygon within which simulated reserves can be placed).
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}

