% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo.rsv}
\alias{demo.rsv}
\title{demo.rsv
An example reserve (e.g., a protected area created for an umbrella species).}
\format{Formal class SpatialPolygons}
\usage{
data(demo.rsv)
}
\description{
demo.rsv
An example reserve (e.g., a protected area created for an umbrella species).
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}

