% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo.sdm}
\alias{demo.sdm}
\title{demo.sdm
An example species distribution model (SDM).
The SDM is coded such that 1 = suitable habitat and 0 = not suitable habitat.}
\format{Formal class RasterLayer}
\usage{
data(demo.sdm)
}
\description{
demo.sdm
An example species distribution model (SDM).
The SDM is coded such that 1 = suitable habitat and 0 = not suitable habitat.
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}

