% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{download}
\alias{download}
\title{Download CHIRPs rasters}
\usage{
download(years, days, output_address, quiet = TRUE)
}
\arguments{
\item{years}{Vector of years}

\item{days}{Vector of days of year}

\item{output_address}{Save location}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and the progress bar.}
}
\description{
Download CHIRPs rasters
}
\examples{
\dontrun{
download(2017:2019, 1:366, "E:/CHIRPS")
}
}
