% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{download_raster}
\alias{download_raster}
\title{Download and unzip the raster}
\usage{
download_raster(url, destination_file)
}
\arguments{
\item{url}{Raster URL}

\item{destination_file}{File address and name, file suffix must by .tif.gz,
the final unzipped file will be with suffix .tif.}
}
\description{
Download and unzip the raster
}
