% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_admin_sf}
\alias{example_admin_sf}
\title{Example sf file.}
\format{
A sf with 2 rows and 4 variables:
}
\source{
\url{https://www.geoboundaries.org/}
}
\usage{
example_admin_sf
}
\description{
Contains boundary data for two admin 1 units. One in Senegal and one in DRC.
These boundaries are from the [GeoBoundaries](https://www.geoboundaries.org/) database.
}
\keyword{datasets}
