% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_daily_rain_raw}
\alias{example_daily_rain_raw}
\title{Example extracted raw rainfall data.}
\format{
A sf with 2 rows and 4 variables:
}
\source{
\url{https://www.chc.ucsb.edu/data/chirps}
}
\usage{
example_daily_rain_raw
}
\description{
Contains rainfall data for two admin 1 units. One in Senegal and one in DRC.
These data are from the [CHIRPS](http://legacy.chg.ucsb.edu/data/chirps/#_Citations) database.
}
\keyword{datasets}
