% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_rain}
\alias{extract_rain}
\title{Extract rainfall from rasters}
\usage{
extract_rain(raster_stack, shape, ...)
}
\arguments{
\item{raster_stack}{Raster stack}

\item{shape}{Spatial polygon (tested with sf)}

\item{...}{Further arguments to  \link[raster]{extract}}
}
\value{
Extracted rainfall in format specified by arguments to \link[raster]{extract}
}
\description{
Extract rainfall from rasters
}
