% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{fit_fourier}
\alias{fit_fourier}
\title{Fit fourier parameters}
\usage{
fit_fourier(rainfall, t, floor = 0.001)
}
\arguments{
\item{rainfall}{Vector of rainfall}

\item{t}{Vector of days (between 1 and 365)}

\item{floor}{Lower bound on rainfall fit}
}
\value{
Model fit
}
\description{
Note without the floor functionality parameters can be more efficiently estimated
with `lm()`.
}
