% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{fit_lm}
\alias{fit_lm}
\title{Linear model fit}
\usage{
fit_lm(rainfall, t)
}
\arguments{
\item{rainfall}{Vector of rainfall}

\item{t}{Vector of days (between 1 and 365)}
}
\description{
Linear model fit
}
