% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_season}
\alias{fit_season}
\title{Fourier fitting function}
\usage{
fit_season(target, t)
}
\arguments{
\item{target}{Vector of rainfall}

\item{t}{Vector of times. A year runs between 0 and 1, so if we have 365 daily
rainfall data points then \code{t = seq(0, 1, length.out = 365)}.
If we had 3 years of daily rainfall points then \code{t = seq(0, 3, length.out = 365 * 3)}.}
}
\value{
Optim output
}
\description{
Fourier fitting function
}
