% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{fourier_model}
\alias{fourier_model}
\title{Estimate the fourier coefficients}
\usage{
fourier_model(rainfall, t)
}
\arguments{
\item{rainfall}{Vector of rainfall}

\item{t}{Vector of times. Must be between 0 and 1.}
}
\value{
Coefficient estimates
}
\description{
Estimate the fourier coefficients
}
