% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{get_urls}
\alias{get_urls}
\title{Get download urls for CHIRPS daily rainfall}
\usage{
get_urls(
  year,
  url_base = "https://data.chc.ucsb.edu/products/CHIRPS-2.0/global_daily/tifs/p25/"
)
}
\arguments{
\item{year}{Year}

\item{url_base}{Base URL}
}
\value{
Daily global rainfall raster urls
}
\description{
Get download urls for CHIRPS daily rainfall
}
