% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getApikey}
\alias{getApikey}
\title{Return a management API Key of the customer}
\usage{
getApikey(serviceProviderId, customerId, keyId)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerId}{The customerID related to this Customer}

\item{keyId}{The API key id}
}
\description{
Return a management API Key of the customer
}
\concept{ManagementAPI}
