% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getCustomer}
\alias{getCustomer}
\title{Return a Customer}
\usage{
getCustomer(serviceProviderId, customerId)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerId}{The customerID related to this Customer}
}
\description{
Return a Customer
}
\concept{ManagementAPI}
