% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getCustomerApikeys}
\alias{getCustomerApikeys}
\title{Return a management API Key of the MspCustomer}
\usage{
getCustomerApikeys(mspId, customerId, keyId)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerId}{The customerID related to this MspCustomer}

\item{keyId}{The API key id}
}
\description{
Return a management API Key of the MspCustomer
}
\concept{ManagementAPI}
