% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DestinationAPIGateway.R
\name{getDestinationList}
\alias{getDestinationList}
\title{Return destination list}
\usage{
getDestinationList(
  destinationListId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{destinationListId}{Destination List ID}

\item{organizationId}{Organization ID}
}
\description{
Return destination list
}
\concept{DestinationAPIGateway}
