% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{getIdentity}
\alias{getIdentity}
\title{Identity}
\usage{
getIdentity(identityid, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{identityid}{the identity id}

\item{organizationId}{the organization id}
}
\description{
Identity
}
\concept{ReportingAPI}
