% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getInternaldomain}
\alias{getInternaldomain}
\title{Return an InternalDomain}
\usage{
getInternaldomain(
  internalDomainId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{internalDomainId}{The internal domain id related to this InternalDomain}

\item{organizationId}{The organization id related to this InternalDomain}
}
\description{
Return an InternalDomain
}
\concept{ManagementAPI}
