% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getInternalnetwork}
\alias{getInternalnetwork}
\title{Return an InternalNetwork}
\usage{
getInternalnetwork(
  internalNetworkId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{internalNetworkId}{The internal network id related to this InternalNetwork}

\item{organizationId}{The organizationID related to this InternalNetwork}
}
\description{
Return an InternalNetwork
}
\concept{ManagementAPI}
