% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getMspCustomers}
\alias{getMspCustomers}
\title{Return a MspCustomer}
\usage{
getMspCustomers(mspId, customerId)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerId}{The customerID related to this MspCustomer}
}
\description{
Return a MspCustomer
}
\concept{ManagementAPI}
