% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getNetwork}
\alias{getNetwork}
\title{Return a Network}
\usage{
getNetwork(networkId, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{networkId}{The network id related to this Network}

\item{organizationId}{The organizationID related to this Network}
}
\description{
Return a Network
}
\concept{ManagementAPI}
