% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmbrellaManagementAPI.R
\name{getNetworkDevice}
\alias{getNetworkDevice}
\title{Retrieve an existing network device}
\usage{
getNetworkDevice(
  originId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{originId}{originID of the identity to remove from the DNS or web policy}

\item{organizationId}{Id of the Organization owning the resources}
}
\description{
Retrieve an existing network device
}
\concept{UmbrellaManagementAPI}
