% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getServiceprovider}
\alias{getServiceprovider}
\title{Return a Service Provider}
\usage{
getServiceprovider(serviceProviderId)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Service Provider}
}
\description{
Return a Service Provider
}
\concept{ManagementAPI}
