% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getSite}
\alias{getSite}
\title{Return a Site}
\usage{
getSite(siteId, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{siteId}{The siteID related to this Site}

\item{organizationId}{The organizationID related to this Site}
}
\description{
Return a Site
}
\concept{ManagementAPI}
