% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{getThreatName}
\alias{getThreatName}
\title{Threat Name}
\usage{
getThreatName(
  threatnameid,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{threatnameid}{the threat-name name}

\item{organizationId}{the organization id}
}
\description{
Threat Name
}
\concept{ReportingAPI}
