% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{getThreatType}
\alias{getThreatType}
\title{Threat Type}
\usage{
getThreatType(
  threattypeid,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{threattypeid}{the threat-type name}

\item{organizationId}{the organization id}
}
\description{
Threat Type
}
\concept{ReportingAPI}
