% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{getTopThreat}
\alias{getTopThreat}
\title{Top Threats (By Type)}
\usage{
getTopThreat(
  type,
  to,
  domains = NULL,
  categories = NULL,
  policycategories = NULL,
  ip = NULL,
  identityids = NULL,
  identitytypes = NULL,
  applicationid = NULL,
  verdict = NULL,
  threats = NULL,
  threattypes = NULL,
  offset = 0,
  limit = 100,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{type}{the type of entries used to calculate the result (dns, proxy, firewall)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{domains}{domain filter, comma delimited}

\item{categories}{categories filter, comma delimited ints}

\item{policycategories}{policy-triggering categories filter, comma delimited ints}

\item{ip}{ip filter}

\item{identityids}{identity filter, comma delimited}

\item{identitytypes}{identity type filter, comma delimited.}

\item{applicationid}{application id filter}

\item{verdict}{verdict filter, comma delimited}

\item{threats}{threat names filter, comma delimited}

\item{threattypes}{threat types filter, comma delimited}

\item{offset}{offset}

\item{limit}{limit}

\item{organizationId}{the organization id}
}
\description{
Gets top threats within timeframe for the given filters.
}
\concept{ReportingAPI}
