% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getUser}
\alias{getUser}
\title{Return a User}
\usage{
getUser(userId, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{userId}{The userID related to this User}

\item{organizationId}{The organizationID related to this User}
}
\description{
Return a User
}
\concept{ManagementAPI}
