% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{getVirtualappliance}
\alias{getVirtualappliance}
\title{Return a VirtualAppliance}
\usage{
getVirtualappliance(
  virtualApplianceId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{virtualApplianceId}{The virtual appliance id related to this VirtualAppliance}

\item{organizationId}{The organizationID related to this VirtualAppliance}
}
\description{
Return a VirtualAppliance
}
\concept{ManagementAPI}
