% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{links}
\alias{links}
\title{Create links}
\usage{
links(core, years, months, days)
}
\arguments{
\item{core}{Core link}

\item{years}{Years}

\item{months}{Months}

\item{days}{Days}
}
\value{
Formtted URL
}
\description{
Create links
}
