% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listAmpRetrospectives}
\alias{listAmpRetrospectives}
\title{Activity AMP Retrospective}
\usage{
listAmpRetrospectives(
  from,
  to,
  ampdisposition = NULL,
  sha256 = NULL,
  offset = 0,
  limit = 100,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{from}{timestamp to filter from (or relative, e.g. -1days)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{ampdisposition}{AMP disposition filter, comma delimited}

\item{sha256}{sha256 filter}

\item{offset}{offset}

\item{limit}{limit}

\item{organizationId}{the organization id}
}
\description{
Gets a list of all AMP retrospective entries within timeframe.
}
\concept{ReportingAPI}
