% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listApikeys}
\alias{listApikeys}
\title{Return a list of management API Keys related to the customer}
\usage{
listApikeys(serviceProviderId, customerId)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerId}{The customerID related to this Customer}
}
\description{
Return a list of management API Keys related to the customer
}
\concept{ManagementAPI}
