% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listApplications}
\alias{listApplications}
\title{Applications}
\usage{
listApplications(
  application = NULL,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{application}{the application name}

\item{organizationId}{the organization id}
}
\description{
Applications
}
\concept{ReportingAPI}
