% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listCategories}
\alias{listCategories}
\title{Categories}
\usage{
listCategories(organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{organizationId}{the organization id}
}
\description{
Categories
}
\concept{ReportingAPI}
