% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listCategoryRequestsByOrg}
\alias{listCategoryRequestsByOrg}
\title{MSP Category Requests by Organization}
\usage{
listCategoryRequestsByOrg(
  msporganizationid,
  from,
  to,
  threats = NULL,
  threattypes = NULL,
  limit = 5000
)
}
\arguments{
\item{msporganizationid}{the provider organization id (multiple orgs can live under a provider)}

\item{from}{timestamp to filter from (or relative, e.g. -1days)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{threats}{threat names filter, comma delimited}

\item{threattypes}{threat types filter, comma delimited}

\item{limit}{limit}
}
\description{
Gets summary counts of all requests for each category per organization within the timeframe
}
\concept{ReportingAPI}
