% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listCustomerApikeys}
\alias{listCustomerApikeys}
\title{Return a list of management API Keys related to the MspCustomer}
\usage{
listCustomerApikeys(mspId, customerId)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerId}{The customerID related to this MspCustomer}
}
\description{
Return a list of management API Keys related to the MspCustomer
}
\concept{ManagementAPI}
