% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listCustomers}
\alias{listCustomers}
\title{Return a list of Customer}
\usage{
listCustomers(serviceProviderId, limit = 100, page = 1)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{limit}{Number of results to return for the given page number.}

\item{page}{Page number of results.}
}
\description{
Return a list of Customer
}
\concept{ManagementAPI}
