% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listDeployments}
\alias{listDeployments}
\title{Deployment Status by Organization}
\usage{
listDeployments(
  msporganizationid,
  from,
  to,
  threats = NULL,
  threattypes = NULL
)
}
\arguments{
\item{msporganizationid}{the provider organization id (multiple orgs can live under a provider)}

\item{from}{timestamp to filter from (or relative, e.g. -1days)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{threats}{threat names filter, comma delimited}

\item{threattypes}{threat types filter, comma delimited}
}
\description{
Gets summary counts of Deployment Status by Organization within timeframe
}
\concept{ReportingAPI}
