% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listDeploymentStatuses}
\alias{listDeploymentStatuses}
\title{Deployment Status}
\usage{
listDeploymentStatuses(
  from,
  to,
  threats = NULL,
  threattypes = NULL,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{from}{timestamp to filter from (or relative, e.g. -1days)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{threats}{threat names filter, comma delimited}

\item{threattypes}{threat types filter, comma delimited}

\item{organizationId}{the organization id}
}
\description{
Gets deployment status within timeframe
}
\concept{ReportingAPI}
