% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DestinationAPIGateway.R
\name{listDestinationLists}
\alias{listDestinationLists}
\title{Retrieve all destination lists of organization}
\usage{
listDestinationLists(organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{organizationId}{Organization ID}
}
\description{
Retrieve all destination lists of organization
}
\concept{DestinationAPIGateway}
