% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listFirewalls}
\alias{listFirewalls}
\title{Activity Firewall}
\usage{
listFirewalls(
  from,
  to,
  ip = NULL,
  ports = NULL,
  identityids = NULL,
  verdict = NULL,
  ruleid = NULL,
  offset = 0,
  limit = 100,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{from}{timestamp to filter from (or relative, e.g. -1days)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{ip}{ip filter}

\item{ports}{81 - port filter, comma delimited ports and ranges}

\item{identityids}{identity filter, comma delimited}

\item{verdict}{verdict filter, comma delimited}

\item{ruleid}{firewall rule ID}

\item{offset}{offset}

\item{limit}{limit}

\item{organizationId}{the organization id}
}
\description{
Gets a list of all Firewall entries within timeframe.
}
\concept{ReportingAPI}
