% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listInternaldomains}
\alias{listInternaldomains}
\title{Return a list of InternalDomains}
\usage{
listInternaldomains(
  limit = 100,
  page = 1,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{limit}{Number of results to return for the given page number.}

\item{page}{Page number of results.}

\item{organizationId}{The organization id related to this InternalDomain}
}
\description{
Return a list of InternalDomains
}
\concept{ManagementAPI}
