% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listInternalnetworkPolicies}
\alias{listInternalnetworkPolicies}
\title{Return a list of Policies for an Internal Network}
\usage{
listInternalnetworkPolicies(
  internalNetworkId,
  type = NULL,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{internalNetworkId}{The internal network id}

\item{type}{}

\item{organizationId}{The organization id}
}
\description{
Return a list of Policies for an Internal Network
}
\concept{ManagementAPI}
