% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listInternalnetworks}
\alias{listInternalnetworks}
\title{Return a list of InternalNetworks}
\usage{
listInternalnetworks(
  limit = 100,
  page = 1,
  name = NULL,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{limit}{Number of results to return for the given page number.}

\item{page}{Page number of results.}

\item{name}{}

\item{organizationId}{The organizationID related to this InternalNetwork}
}
\description{
Return a list of InternalNetworks
}
\concept{ManagementAPI}
