% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listIps}
\alias{listIps}
\title{Activity IP}
\usage{
listIps(
  from,
  to,
  categories = NULL,
  ip = NULL,
  ports = NULL,
  identityids = NULL,
  identitytypes = NULL,
  verdict = NULL,
  offset = 0,
  limit = 5000,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{from}{timestamp to filter from (or relative, e.g. -1days)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{categories}{categories filter, comma delimited ints}

\item{ip}{ip filter}

\item{ports}{81 - port filter, comma delimited ports and ranges}

\item{identityids}{identity filter, comma delimited}

\item{identitytypes}{identity type filter, comma delimited.}

\item{verdict}{verdict filter, comma delimited}

\item{offset}{offset}

\item{limit}{limit}

\item{organizationId}{the organization id}
}
\description{
Gets a list of all IP entries within timeframe.
}
\concept{ReportingAPI}
