% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listMspCustomers}
\alias{listMspCustomers}
\title{Return a list of MspCustomer}
\usage{
listMspCustomers(mspId, limit = 100, page = 1)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{limit}{Number of results to return for the given page number.}

\item{page}{Page number of results.}
}
\description{
Return a list of MspCustomer
}
\concept{ManagementAPI}
