% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmbrellaManagementAPI.R
\name{listNetworkDevices}
\alias{listNetworkDevices}
\title{Retrieve all network devices for the current organization}
\usage{
listNetworkDevices(organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{organizationId}{Id of the Organization owning the resources}
}
\description{
Retrieve all network devices for the current organization
}
\concept{UmbrellaManagementAPI}
