% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmbrellaManagementAPI.R
\name{listOrganizationPolicies}
\alias{listOrganizationPolicies}
\title{List Policies}
\usage{
listOrganizationPolicies(
  page = 1,
  limit = 100,
  type = "DNS",
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{page}{Start of the page}

\item{limit}{Number of items to return}

\item{type}{The type of policies to be returned (either DNS or web)}

\item{organizationId}{Id of the Organization owning the resources}
}
\description{
List DNS and web policies, if no filters are supplied then only DNS policies accessible with the current credentials will be returned.
}
\concept{UmbrellaManagementAPI}
