% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmbrellaManagementAPI.R
\name{listOrganizations}
\alias{listOrganizations}
\title{List organizations}
\usage{
listOrganizations()
}
\description{
List Organizations available to the current API client. Currently this will return a single Organization. In the future it may be possible to access multiple Umbrella Organizations with a single set of credentials.
}
\concept{UmbrellaManagementAPI}
