% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listPolicies}
\alias{listPolicies}
\title{Return a list of Policies for a Network}
\usage{
listPolicies(
  networkId,
  type = NULL,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{networkId}{The network id}

\item{type}{}

\item{organizationId}{The organization id}
}
\description{
Return a list of Policies for a Network
}
\concept{ManagementAPI}
