% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmbrellaManagementAPI.R
\name{listPoliciesForDevice}
\alias{listPoliciesForDevice}
\title{List the Policies associated with a Network Device}
\usage{
listPoliciesForDevice(
  originId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{originId}{originID of the identity to remove from the DNS or web policy}

\item{organizationId}{Id of the Organization owning the resources}
}
\description{
List the Policies associated with a Network Device
}
\concept{UmbrellaManagementAPI}
