% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listProviderCategories}
\alias{listProviderCategories}
\title{MSP Categories}
\usage{
listProviderCategories(msporganizationid)
}
\arguments{
\item{msporganizationid}{the provider organization id (multiple orgs can live under a provider)}
}
\description{
MSP Categories
}
\concept{ReportingAPI}
