% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listProviderRequestsByTimerange}
\alias{listProviderRequestsByTimerange}
\title{MSP Requests by Timerange (ALL)}
\usage{
listProviderRequestsByTimerange(
  msporganizationid,
  from,
  to,
  domains = NULL,
  urls = NULL,
  categories = NULL,
  policycategories = NULL,
  ip = NULL,
  identityids = NULL,
  identitytypes = NULL,
  applicationid = NULL,
  verdict = NULL,
  threats = NULL,
  threattypes = NULL,
  offset = 0,
  limit = 5000
)
}
\arguments{
\item{msporganizationid}{the provider organization id (multiple orgs can live under a provider)}

\item{from}{timestamp to filter from (or relative, e.g. -1days)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{domains}{domain filter, comma delimited}

\item{urls}{url filter, comma delimited}

\item{categories}{categories filter, comma delimited ints}

\item{policycategories}{policy-triggering categories filter, comma delimited ints}

\item{ip}{ip filter}

\item{identityids}{identity filter, comma delimited}

\item{identitytypes}{identity type filter, comma delimited.}

\item{applicationid}{application id filter}

\item{verdict}{verdict filter, comma delimited}

\item{threats}{threat names filter, comma delimited}

\item{threattypes}{threat types filter, comma delimited}

\item{offset}{offset}

\item{limit}{limit}
}
\description{
Gets activity volume within timeframe
}
\concept{ReportingAPI}
