% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listRequestsByDestination}
\alias{listRequestsByDestination}
\title{MSP Requests by Destination}
\usage{
listRequestsByDestination(
  msporganizationid,
  from,
  to,
  threats = NULL,
  threattypes = NULL,
  limit = 100
)
}
\arguments{
\item{msporganizationid}{the provider organization id (multiple orgs can live under a provider)}

\item{from}{timestamp to filter from (or relative, e.g. -1days)}

\item{to}{timestamp to filter to  (or relative, e.g. now)}

\item{threats}{threat names filter, comma delimited}

\item{threattypes}{threat types filter, comma delimited}

\item{limit}{limit}
}
\description{
Gets summary counts of all requests within timeframe
}
\concept{ReportingAPI}
