% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listRoamingcomputers}
\alias{listRoamingcomputers}
\title{Return a list of RoamingComputers}
\usage{
listRoamingcomputers(
  limit = 100,
  page = 1,
  name = NULL,
  status = NULL,
  lastSyncBefore = NULL,
  lastSyncAfter = NULL,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{limit}{Number of results to return for the given page number.}

\item{page}{Page number of results.}

\item{name}{}

\item{status}{}

\item{lastSyncBefore}{}

\item{lastSyncAfter}{}

\item{organizationId}{The organizationID related to this RoamingComputer}
}
\description{
Return a list of RoamingComputers
}
\concept{ManagementAPI}
