% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listRoles}
\alias{listRoles}
\title{Return a list of Roles}
\usage{
listRoles(organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{organizationId}{The organizationID related to this Role}
}
\description{
Return a list of Roles
}
\concept{ManagementAPI}
