% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listThreatNames}
\alias{listThreatNames}
\title{Threat Names}
\usage{
listThreatNames(organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{organizationId}{the organization id}
}
\description{
Threat Names
}
\concept{ReportingAPI}
