% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingAPI.R
\name{listThreatTypes}
\alias{listThreatTypes}
\title{Threat Types}
\usage{
listThreatTypes(organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{organizationId}{the organization id}
}
\description{
Threat Types
}
\concept{ReportingAPI}
