% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listUsers}
\alias{listUsers}
\title{Return a list of Users}
\usage{
listUsers(
  limit = 100,
  page = 1,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{limit}{Number of results to return for the given page number.}

\item{page}{Page number of results.}

\item{organizationId}{The organizationID related to this User}
}
\description{
Return a list of Users
}
\concept{ManagementAPI}
