% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{listVirtualappliances}
\alias{listVirtualappliances}
\title{Return a list of VirtualAppliances}
\usage{
listVirtualappliances(
  limit = 100,
  page = 1,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{limit}{Number of results to return for the given page number.}

\item{page}{Page number of results.}

\item{organizationId}{The organizationID related to this VirtualAppliance}
}
\description{
Return a list of VirtualAppliances
}
\concept{ManagementAPI}
