% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifyApikey}
\alias{modifyApikey}
\title{Update to generate a fresh management API key-secret for the customer}
\usage{
modifyApikey(serviceProviderId, customerId, keyId)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerId}{The customerID related to this Customer}

\item{keyId}{The existing API key id}
}
\description{
Update to generate a fresh management API key-secret for the customer
}
\concept{ManagementAPI}
