% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifyCustomer}
\alias{modifyCustomer}
\title{Update a Customer}
\usage{
modifyCustomer(
  serviceProviderId,
  customerId,
  customerName,
  seats,
  streetAddress,
  streetAddress2 = NULL,
  city,
  state,
  countryCode,
  zipCode = NULL,
  packageId,
  dealId = NULL,
  adminEmails,
  ccwDealOwnerEmails = NULL
)
}
\arguments{
\item{serviceProviderId}{The ServiceProviderID related to this Customer}

\item{customerId}{The customerID related to this Customer}

\item{customerName}{customer organization name}

\item{seats}{number of users}

\item{streetAddress}{street address}

\item{streetAddress2}{street address2}

\item{city}{city}

\item{state}{state}

\item{countryCode}{country code}

\item{zipCode}{zip code}

\item{packageId}{package id}

\item{dealId}{deal id}

\item{adminEmails}{array of admin emails}

\item{ccwDealOwnerEmails}{array of CCW deal owner emails}
}
\description{
Update a Customer
}
\concept{ManagementAPI}
