% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifyCustomerApikeys}
\alias{modifyCustomerApikeys}
\title{Update to generate a fresh management API key-secret for the MspCustomer}
\usage{
modifyCustomerApikeys(mspId, customerId, keyId)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerId}{The customerID related to this MspCustomer}

\item{keyId}{The existing API key id}
}
\description{
Update to generate a fresh management API key-secret for the MspCustomer
}
\concept{ManagementAPI}
