% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DestinationAPIGateway.R
\name{modifyDestinationList}
\alias{modifyDestinationList}
\title{Rename destination list}
\usage{
modifyDestinationList(
  destinationListId,
  body = DestinationListPatch(),
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{destinationListId}{Destination List ID}

\item{organizationId}{Organization ID}
}
\description{
Rename destination list
}
\concept{DestinationAPIGateway}
