% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifyInternalnetwork}
\alias{modifyInternalnetwork}
\title{Update an InternalNetwork}
\usage{
modifyInternalnetwork(
  internalNetworkId,
  name,
  ipAddress,
  prefixLength,
  siteId = NULL,
  networkId = NULL,
  tunnelId = NULL,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{internalNetworkId}{The internal network id related to this InternalNetwork}

\item{name}{internal network label}

\item{ipAddress}{internal network ipv4 address}

\item{prefixLength}{prefix length}

\item{siteId}{For DNS policies: ID of Site to be associated with Internal Network. (Provide only one of siteId, networkId, or tunnelId)}

\item{networkId}{For Web policies via proxy chaining: ID of Network to be associated with Internal Network. (Provide only one of siteId, networkId, or tunnelId)}

\item{tunnelId}{For Web policies via IPsec tunnel: ID of Tunnel to be associated with Internal Network. (Provide only one of siteId, networkId, or tunnelId)}

\item{organizationId}{The organizationID related to this InternalNetwork}
}
\description{
Update an InternalNetwork
}
\concept{ManagementAPI}
