% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifyMspCustomers}
\alias{modifyMspCustomers}
\title{Update a MspCustomer}
\usage{
modifyMspCustomers(mspId, customerId, customerName, seats)
}
\arguments{
\item{mspId}{The mspID related to this MspCustomer}

\item{customerId}{The customerID related to this MspCustomer}

\item{customerName}{customer organization name}

\item{seats}{number of users}
}
\description{
Update a MspCustomer
}
\concept{ManagementAPI}
