% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifyNetwork}
\alias{modifyNetwork}
\title{Update a Network}
\usage{
modifyNetwork(
  networkId,
  name,
  ipAddress = NULL,
  prefixLength = NULL,
  isDynamic,
  status,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{networkId}{The network id related to this Network}

\item{name}{network name}

\item{ipAddress}{network ip address}

\item{prefixLength}{prefix length, must be greater than 28}

\item{isDynamic}{is dynamic ip}

\item{status}{status}

\item{organizationId}{The organizationID related to this Network}
}
\description{
Update a Network
}
\concept{ManagementAPI}
