% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmbrellaManagementAPI.R
\name{modifyNetworkDevice}
\alias{modifyNetworkDevice}
\title{Update a network device}
\usage{
modifyNetworkDevice(
  originId,
  name,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{originId}{originID of the identity to remove from the DNS or web policy}

\item{name}{}

\item{organizationId}{Id of the Organization owning the resources}
}
\description{
Update a network device
}
\concept{UmbrellaManagementAPI}
