% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifyRoamingcomputer}
\alias{modifyRoamingcomputer}
\title{Update a RoamingComputer}
\usage{
modifyRoamingcomputer(
  roamingComputerId,
  name,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{roamingComputerId}{The roamingComputerID related to this RoamingComputer}

\item{name}{roaming computer name}

\item{organizationId}{The organizationID related to this RoamingComputer}
}
\description{
Update a RoamingComputer
}
\concept{ManagementAPI}
