% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifySite}
\alias{modifySite}
\title{Update a Site}
\usage{
modifySite(siteId, name, organizationId = Sys.getenv("umbrellaOrganizationId"))
}
\arguments{
\item{siteId}{The siteID related to this Site}

\item{name}{site name}

\item{organizationId}{The organizationID related to this Site}
}
\description{
Update a Site
}
\concept{ManagementAPI}
