% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagementAPI.R
\name{modifyVirtualappliance}
\alias{modifyVirtualappliance}
\title{Update a VirtualAppliance}
\usage{
modifyVirtualappliance(
  virtualApplianceId,
  siteId,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{virtualApplianceId}{The virtual appliance id related to this VirtualAppliance}

\item{siteId}{site id}

\item{organizationId}{The organizationID related to this VirtualAppliance}
}
\description{
Update a VirtualAppliance
}
\concept{ManagementAPI}
