% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{objective}
\alias{objective}
\title{Objective function for fitting}
\usage{
objective(coef, t, floor, rainfall)
}
\arguments{
\item{coef}{Fourier coefficients}

\item{t}{Vector of timesteps (between 0 and 1)}

\item{floor}{Lower bound on rainfall fit}

\item{rainfall}{Vector of rainfall}
}
\value{
Sum of squared differences between profile and rainfall data
}
\description{
Objective function for fitting
}
