% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process rainfall data}
\usage{
process(gee_output, ...)
}
\arguments{
\item{gee_output}{Output from gee call using \code{\link{pull_daily_rainfall}}}

\item{...}{Grouping variables}
}
\value{
Model, coefficients, representative seasonalilty profile and rainfall data per group
}
\description{
Process rainfall data
}
