% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull_daily_rainfall}
\alias{pull_daily_rainfall}
\title{Extract mean daily precipitation}
\usage{
pull_daily_rainfall(sf, start_date, end_date)
}
\arguments{
\item{sf}{A simple features object. See \code{\link[sf]{sf}} for more details.}

\item{start_date}{First extraction time point. A string with format: "yyyy-mm-dd".}

\item{end_date}{Last extraction time point. A string with format: "yyyy-mm-dd".}
}
\value{
Matrix
}
\description{
Extract mean daily precipitation
}
