% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Example rainfall data.}
\format{
An object of class \code{data.frame} with 365 rows and 2 columns.
}
\usage{
rainfall
}
\description{
Example rainfall data.
}
\keyword{datasets}
