% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{s_a}
\alias{s_a}
\title{Fourier transform internal cos}
\usage{
s_a(a, t, i)
}
\arguments{
\item{a}{Parameter}

\item{t}{Vector of times}

\item{i}{Level}
}
\value{
Fourier component
}
\description{
Fourier transform internal cos
}
