% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{s_b}
\alias{s_b}
\title{Fourier transform internal sin}
\usage{
s_b(b, t, i)
}
\arguments{
\item{b}{Parameter}

\item{t}{Vector of times}

\item{i}{Level}
}
\value{
Fourier component
}
\description{
Fourier transform internal sin
}
