% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{season_optim}
\alias{season_optim}
\title{Fourier fitting function internal}
\usage{
season_optim(par, target, t, ss = TRUE)
}
\arguments{
\item{par}{Vector of parameters}

\item{target}{Vector of rainfall}

\item{t}{Vector of times. A year runs between 0 and 1, so if we have 365 daily
rainfall data points then \code{t = seq(0, 1, length.out = 365)}.
If we had 3 years of daily rainfall points then \code{t = seq(0, 3, length.out = 365 * 3)}.}

\item{ss}{Return sum of squares or prediction}
}
\value{
if \code{ss = TRUE} Sum of squared differences else prediction
}
\description{
Fourier fitting function internal
}
