% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{seasonality}
\alias{seasonality}
\title{Seasonality curve}
\usage{
seasonality(par, t = seq(0, 1, length.out = 365))
}
\arguments{
\item{par}{Parameter vector: a0, a1, b2, a2, b2, a3, b3}

\item{t}{Vector of times. One year runs between 0 and 1.}
}
\value{
seasonality curve
}
\description{
Predicts a seasonal curve given fourier transform parameters
}
