% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simReserve.R
\name{simReserve}
\alias{simReserve}
\title{Simulate randomly sited reserve}
\usage{
simReserve(target.poly, buff.width, total.area, wiggle)
}
\arguments{
\item{target.poly}{A polygon within which all simulated reserves are to be contained.  The availability mask.}

\item{buff.width}{The radius of the randomly sited circles to be used.}

\item{total.area}{The target size of the simulated reserve.}

\item{wiggle}{How much bigger/smaller can the simulated reserves be than \code{tota.area}
(in same units as \code{tota.area})}
}
\value{
Object of class SpatialPolygons, the simulated reserve.
}
\description{
This function is used to generate a randomly sited reserve, more specifically,
an aggregate of multiple reserve units (distinct polygons).
}
\details{
The reserve is built by randomly siting circular polygons of the specified size.  The polygons are allowed to overlap.
Circles are iteratively added until the specified area is reached, then a smaller circle is trimmed out if
the simulated reserve overshot the specified area.  The user can specify a tolerance for how much bigger/smaller
than the specified area is acceptable.  This function has been further developed in the ecoinfo package, contact
the author for details.
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}
