% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumRaster.R
\name{sumRaster}
\alias{sumRaster}
\title{Sum of raster within polygon}
\usage{
sumRaster(rast, poly)
}
\arguments{
\item{rast}{A raster object of class RasterLayer.}

\item{poly}{A spatial polygon object of class SpatialPolygons.}
}
\value{
Numeric, sum of raster cells within a polygon.
}
\description{
Returns the sum of raster cells within a polygon.  When the raster is coded as a typical SDM (where 1 = suitable habitat
and 0 = non-suitable habitat), returns the number of suitable cells within the polygon.
}
\details{
The same process can be done with \code{raster::extract}; however, this
function (uses \code{raster::mask} and \code{raster::cellStats}) performed faster in a local test using large raster objects.
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}

