% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{umbrella}
\alias{umbrella}
\alias{umbrella-package}
\title{umbrella:  Evaluating Conservation Reserves}
\description{
This package contains functions (and example data) that aid in evaluating the performance of a conservation reserve in
protecting the suitable habitat of species of conservation interest.  This package was designed specifically to assess
conservation reserves created for an umbrella species--hence the name.
}
\details{
\bold{version 0.1.0}:  Contains two functions (\code{sumRaster} and \code{simReserve}) and three example
datasets (\code{demo.sdm}, \code{demo.rsv}, and \code{demo.msk}).  Demo of package submitted as supplement to
published manuscript.

Bug reports or suggestions for the umbrella package should be directed to <jason.d.carlisle@gmail.com>
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}

