% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{urls}
\alias{urls}
\title{Create vector of URLs for CHIRPS daily rainfall download}
\usage{
urls(year)
}
\arguments{
\item{year}{A year}
}
\value{
Vector of URLs
}
\description{
Create vector of URLs for CHIRPS daily rainfall download
}
