% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{data}
\name{demo.msk}
\alias{demo.msk}
\title{demo.msk}
\format{Formal class SpatialPolygons}
\usage{
data(demo.msk)
}
\description{
An example availability mask (i.e., the polygon within which simulated reserves can be placed).
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}
