% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{data}
\name{demo.rsv}
\alias{demo.rsv}
\title{demo.rsv}
\format{Formal class SpatialPolygons}
\usage{
data(demo.rsv)
}
\description{
An example reserve (e.g., a protected area created for an umbrella species).
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}
