% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{fourier_predict}
\alias{fourier_predict}
\title{Estimate the fourier coefficients}
\usage{
fourier_predict(coef, t, floor)
}
\arguments{
\item{coef}{Fourier coefficients}

\item{t}{Vector of days (between 1 and 365)}

\item{floor}{Lower bound on rainfall fit}
}
\description{
Estimate the fourier coefficients
}
