% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DestinationAPIGateway.R
\name{listDestinations}
\alias{listDestinations}
\title{Get list of destinations related to destination list}
\usage{
listDestinations(
  destinationListId,
  page = 1,
  limit = 100,
  organizationId = Sys.getenv("umbrellaOrganizationId")
)
}
\arguments{
\item{destinationListId}{Destination List ID}

\item{page}{Pagination}

\item{limit}{Limit for page}

\item{organizationId}{Organization ID}
}
\description{
Get list of destinations related to destination list
}
\concept{DestinationAPIGateway}
